+++ EppoEngine/Source/Renderer/Buffer/IndexBuffer.h
@@ -1,24 +1,14 @@
 #pragma once
 
-#include "Renderer/Allocator.h"
-
 namespace Eppo
 {
 	class IndexBuffer
 	{
 	public:
-		IndexBuffer(void* data, uint32_t size);
-		~IndexBuffer();
-		IndexBuffer(const IndexBuffer&) = delete;
-		IndexBuffer& operator=(const IndexBuffer&) = delete;
+		virtual ~IndexBuffer() {}
 
-		VkBuffer GetBuffer() const { return m_Buffer; }
-		uint32_t GetIndexCount() const { return m_Size / sizeof(uint32_t); }
+		virtual uint32_t GetIndexCount() const = 0;
 
-	private:
-		uint32_t m_Size;
-
-		VkBuffer m_Buffer;
-		VmaAllocation m_Allocation;
+		static Ref<IndexBuffer> Create(void* data, uint32_t size);
 	};
 }
